+++ src/theme/components/button.ts
@@ -6,8 +6,8 @@ const baseStyle = {
   cursor: "pointer",
   color: "ui.white",
   justifyContent: "center",
-  py: "xs",
-  px: "s",
+  py: "inset.narrow",
+  px: "inset.default",
   textDecoration: "none",
   wordWrap: "normal",
   fontWeight: "button.default",
@@ -69,9 +69,8 @@ const pill = {
   borderColor: "ui.gray.light-cool",
   color: "inherit",
   borderRadius: "pill",
-  py: "xs",
-  paddingInlineStart: "m",
-  paddingInlineEnd: "m",
+  py: "inset.narrow",
+  px: "inset.wide",
   fontSize: "button.default",
   _hover: {
     bg: "ui.gray.xx-light-cool",
@@ -89,8 +88,8 @@ const iconOnly = {
   _hover: {
     bg: "ui.gray.xx-light-cool",
   },
-  paddingInlineStart: "xs",
-  paddingInlineEnd: "xs",
+  paddingInlineStart: "inset.narrow",
+  paddingInlineEnd: "inset.narrow",
 };
 const callout = {
   bg: "brand.primary",