+++ .storybook/main.js
@@ -30,6 +30,7 @@ module.exports = {
     "@storybook/addon-controls",
     "@storybook/addon-queryparams",
     "@storybook/addon-jest",
+    "@chakra-ui/storybook-addon",
   ],
   typescript: {
     // Type-check stories during Storybook build.