+++ CHANGELOG.md
@@ -41,7 +41,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Refactors the DS `RadioGroup` component so it internally implements Chakra's `RadioGroup` component rather than the `useRadioGroup` hook. The "uncontrolled" version of Chakra's `RadioGroup` is not working and will be investigated in the future. It is recommended to use the controlled component pattern.
 - Updates the `Fieldset` component to render the "Optional"/"Required" text in the `legend` element as pseudo CSS in the `::after` rule.
 - Passes the `isRequired` prop in the `RadioGroup` and `CheckboxGroup` to the `Fieldset` wrapper component.
-- Updates the `focus`  styles for the `Radio` and `TextInput` components.
+- Updates the `focus` styles for the `Radio` and `TextInput` components.
 
 ### Removals