+++ CHANGELOG.md
@@ -47,7 +47,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the `focus` styles for the `Radio`, `Select` and `TextInput` components.
 - Updates the references of the `<dl>` element from "Definition" to "Description", as that's the official name in HTML5. This affects the `List` element and its `ListTypes.Description` enum value.
 - Updates styles for `React Datepicker`'s calendar popup in the `DatePicker` component.
-- Updates storybook URL from /nypl-design-system/storybook-static/ to /nypl-design-system/reservoir/v0_26.
+- Updates storybook URL from /storybook-static/ to /reservoir/v0_26.
 
 ### Removals