+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds the `isAlignedRightActions` prop to the `Card` component to render `CardActions` components to the right of the main content area. This only works for the `Card`'s row layout.
 - Adds styles to target native HTML basic elements inside the `StructuredContent` component.
 - Adds `Reservoir` branding to Storybook.
+- Adds `@chakra-ui/storybook-addon` so stories automatically pick up the Chakra-based NYPL theme.
 
 ### Changes
 
@@ -51,6 +52,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates storybook URL from /storybook-static/ to /reservoir/v0_26.
 - Updates all references to "design system" in the Storybook documentation to include "Reservoir" in the name of the design system.
 - Updates language on the `Welcome` page.
+- Updates Chakra packages `@chakra-ui/react` and `@chakra-ui/system`.
 
 ### Removals